Cleaned out all of the unused dependencies and separated functionality from view file and put functionality into flowBuilder.js
